third_party/blink/web_tests/platform/generic/fast/overflow/child-100percent-height-inside-fixed-container-with-overflow-auto-expected.txt

Testcase for bug https://bugs.webkit.org/show_bug.cgi?id=11355. When a container having overflow:auto has a horizontal scrollbar, the scrollbar is to be placed between the inner border edge and the outer padding edge. Thus the content height of a child inside the container must not include the height of the horizontal scollbars.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

The height of the inner element box should be 100% of the containers height minus the height of horizontal scrollbar. There should be no vertical scrollable content in the container

PASS document.getElementById("container").scrollHeight == document.getElementById("container").clientHeight is true
Container height = Inner Box height + scrollbar height
PASS document.getElementById("container").offsetHeight > document.getElementById("innerBox").offsetHeight is true
PASS successfullyParsed is true

TEST COMPLETE
